# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for GoPOS Backend",
        "title": "GoPOS API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/users": {
            "get": {
                "description": "Get a full list of registered accounts ordered by their unique database ID. Restricted to admin access only.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users Management"
                ],
                "summary": "Retrieve all users",
                "responses": {
                    "200": {
                        "description": "Successfully fetched all system users",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Register a new cashier (kasir) or administrator account. Restricted to admin access only.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users Management"
                ],
                "summary": "Create a new user account",
                "parameters": [
                    {
                        "description": "User Creation Payload Data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateUserInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User successfully created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid input request format or invalid role assignment",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "409": {
                        "description": "Email address is already registered",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error during password encryption",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/users/{id}/activate": {
            "put": {
                "description": "Restore account system operations by resetting the active status flag back to true. Restricted to admin access only.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users Management"
                ],
                "summary": "Re-activate a suspended user account",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Target Numeric User Database ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User profile successfully re-activated",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "No user entry found matching the given target ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Database connection error updating row state",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/users/{id}/deactivate": {
            "put": {
                "description": "Set user active status flag to false to temporarily revoke platform access. Self-deactivation is prohibited. Restricted to admin access only.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users Management"
                ],
                "summary": "Deactivate a user account (Soft Disable)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Target Numeric User Database ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User profile successfully disabled",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Action rejected because an administrator cannot suspend their own session account",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "No user entry found matching the given target ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Database connection error modifying data",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.CreateUserInput": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password",
                "role"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}